Редьюср разбит на слайсы, добавлены селекторы, рефаторинг компонентов под новый редьюсер и селекторы
