+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import App from './components/app/app';
-import ErrorMessage from './components/error-message/error-message';
+// import ErrorMessage from './components/error-message/error-message';
 import { CITIES_NAMES } from './const';
 import { checkAuthAction, fetchOffersAction } from './store/api-actions';
 import { store } from './store/store';
@@ -17,7 +17,7 @@ const root = ReactDOM.createRoot(
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <ErrorMessage />
+      {/* <ErrorMessage /> */}
       <App citiesNames={CITIES_NAMES} />
     </Provider>
   </React.StrictMode>
