+++ src/services/process-error-handle.ts
@@ -1,11 +1,11 @@
+// // import { store } from '../store/store';
+// import { TIMEOUT_SHOW_ERROR } from '../const';
+// import { setError } from '../store/action';
 // import { store } from '../store/store';
-import { TIMEOUT_SHOW_ERROR } from '../const';
-import { setError } from '../store/action';
-import { store } from '../store/store';
 
-export const processErrorHandle = (message: string): void => {
-  store.dispatch(setError(message));
-  setTimeout(() => {
-    store.dispatch(setError(null));
-  }, TIMEOUT_SHOW_ERROR);
-};
+// export const processErrorHandle = (message: string): void => {
+//   store.dispatch(setError(message));
+//   setTimeout(() => {
+//     store.dispatch(setError(null));
+//   }, TIMEOUT_SHOW_ERROR);
+// };
