+++ src/store/store.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer';
 import { createAPI } from '../services/api';
+import { rootReducer } from './root-reducer';
 
 export type Store = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
@@ -8,7 +8,7 @@ export type AppDispatch = typeof store.dispatch;
 export const api = createAPI();
 
 export const store = configureStore({
-  reducer,
+  reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
